phase6: refactoring
