+++ tests/phase2/test_search_issue.py
@@ -1,6 +1,5 @@
 from http import HTTPStatus
 
-from src.common import get_time_stamp
 from src.jira_api import ApiIssue, ApiSearch
 
 from src.constants import *
@@ -17,19 +16,8 @@ def test_search_one_issue(jira_tests_fixture, api_issue_fixture, created_dummy_i
     assert created_dummy_issue.get_summary() == r.json()["issues"][0]["fields"]["summary"]
 
 
-def test_search_multiple_issues(jira_tests_fixture, api_issue_fixture, created_issues):
+def test_search_multiple_issues(jira_tests_fixture, api_issue_fixture, created_issues, created_issues_summaries):
     api_session, jira_session = jira_tests_fixture
-
-    # create several issues
-    created_issues_summaries = []
-    for issue_num in range(MAX_ISSUES_TO_CREATE):
-        summary = "Oleg " + get_time_stamp()
-        api_issue = ApiIssue(jira_session, PROJECT_KEY, BUG_ISSUE_TYPE_KEY, summary)
-        api_issue_fixture.set(api_issue)
-        r = api_issue.create_issue()
-        assert HTTPStatus.CREATED == r.status_code
-        created_issues.append(r.json()["id"])
-        created_issues_summaries.append(summary)
 
     # search for multiple issues
     jql_search_prepared = JQL_QUERY + " OR summary ~ ".join(['"' + x + '"' for x in created_issues_summaries])
