+++ tests/phase2/test_update.py
@@ -1,12 +1,7 @@
 from http import HTTPStatus
 
-from src.common import get_time_stamp
-from src.jira_api import ApiIssue
-from src.constants import *
-
 
 def test_update_issue(api_issue_fixture, created_issues, created_dummy_issue):
-
     # update issue
     created_dummy_issue.set_summary("Hello")
     created_dummy_issue.set_assignee("Oleg_Moskalyov")
@@ -20,4 +15,3 @@ def test_update_issue(api_issue_fixture, created_issues, created_dummy_issue):
     assert r.json()["fields"]["summary"] == created_dummy_issue.get_summary()
     assert r.json()["fields"]["assignee"]["name"] == created_dummy_issue.get_assignee()
     assert r.json()["fields"]["priority"]["name"] == created_dummy_issue.get_priority()
-
